Re-architecture Phase 3: factor entity/dependency/comparison logic into core
org.finos.legend.sdlc.core.entity gains EntityAccessOperations,
EntityModificationOperations, and EntityProjectFile - the entity read/write
logic previously duplicated between GitLabEntityApi and FileSystemEntityApi,
written once over FileAccessContext + ProjectStructure. core.dependency gains
the transitive upstream dependency walk (parameterized over a configuration
resolver) and core.comparison gains the entity-diff assembly (over
backend-neutral FileDiffs) plus a generic two-context walking comparison.
GitLab and FS api classes delegate and shrink but do not move; FS keeps its
git-tree-walk enumeration and quirks, GitLab keeps its native compare and its
buildException error translation. Core throws LegendSDLCException;
BaseGitLabApi's exception pass-through widened to the base type (first part of
the Phase 2 catch-site audit). One deliberate error-path message drift and the
delegation boundaries are documented in the worklog; characterization pins
updated accordingly (thrown-type assertions widened to the base per the
Phase 2 precedent).
