+++ legend-sdlc-core/pom.xml
@@ -54,6 +54,11 @@
             <artifactId>eclipse-collections</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+        </dependency>
+
         <!-- TEST -->
         <dependency>
             <groupId>junit</groupId>
