+++ legend-sdlc-server-fs/pom.xml
@@ -155,10 +155,6 @@
             <artifactId>eclipse-collections-api</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.eclipse.collections</groupId>
-            <artifactId>eclipse-collections</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.eclipse.jgit</groupId>
             <artifactId>org.eclipse.jgit</artifactId>
             <version>6.3.0.202209071007-r</version>
