+++ legend-sdlc-server/src/main/java/org/finos/legend/sdlc/server/gitlab/api/GitLabComparisonApi.java
@@ -14,18 +14,16 @@
 
 package org.finos.legend.sdlc.server.gitlab.api;
 
-import org.eclipse.collections.api.factory.Lists;
+import org.eclipse.collections.impl.utility.ListIterate;
+import org.finos.legend.sdlc.core.comparison.ComparisonOperations;
+import org.finos.legend.sdlc.core.comparison.FileDiff;
 import org.finos.legend.sdlc.domain.model.comparison.Comparison;
-import org.finos.legend.sdlc.domain.model.comparison.EntityDiff;
-import org.finos.legend.sdlc.domain.model.entity.change.EntityChangeType;
 import org.finos.legend.sdlc.server.domain.api.comparison.ComparisonApi;
 import org.finos.legend.sdlc.server.domain.api.workspace.WorkspaceSpecification;
 import org.finos.legend.sdlc.server.error.LegendSDLCServerException;
 import org.finos.legend.sdlc.server.gitlab.GitLabConfiguration;
 import org.finos.legend.sdlc.server.gitlab.GitLabProjectId;
 import org.finos.legend.sdlc.server.gitlab.auth.GitLabUserContext;
-import org.finos.legend.sdlc.project.structure.EntitySourceDirectory;
-import org.finos.legend.sdlc.project.files.ProjectPaths;
 import org.finos.legend.sdlc.project.structure.ProjectStructure;
 import org.finos.legend.sdlc.server.tools.BackgroundTaskProcessor;
 import org.gitlab4j.api.GitLabApi;
@@ -39,7 +37,6 @@ import org.gitlab4j.api.models.MergeRequest;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicBoolean;
 import javax.inject.Inject;
 import javax.ws.rs.core.Response;
 
@@ -246,104 +243,12 @@ public class GitLabComparisonApi extends GitLabApiWithFileAccess implements Comp
         {
             throw new LegendSDLCServerException("Unexpected Comparison Result: toRevisionId does not match expected. Expected: " + toRevisionId + ", Actual: " + comparisonResultCommit.getId());
         }
-        return newComparison(fromRevisionId, toRevisionId, comparisonResult.getDiffs(), fromProjectStructure, toProjectStructure);
+        List<FileDiff> fileDiffs = ListIterate.collect(comparisonResult.getDiffs(), GitLabComparisonApi::toFileDiff);
+        return ComparisonOperations.newComparison(fromRevisionId, toRevisionId, fileDiffs, fromProjectStructure, toProjectStructure);
     }
 
-    private static Comparison newComparison(String fromRevisionId, String toRevisionId, List<Diff> deltas, ProjectStructure fromProjectStructure, ProjectStructure toProjectStructure)
+    private static FileDiff toFileDiff(Diff diff)
     {
-        List<EntityDiff> entityDiffs = Lists.mutable.empty();
-        AtomicBoolean isProjectConfigurationUpdated = new AtomicBoolean(false);
-        deltas.forEach(diff ->
-        {
-            // File changes can be of three types:
-            // 1. entity file changes - which we should handle without any problem
-            // 2. project configuration changes - which we will just capture by a boolean flag to indicate if there are any changes
-            // 3. other files: e.g. users can go in and modify pom.xml, add some non-entity files, etc. we DO NOT keep track of these
-            String oldPath = ProjectPaths.canonicalizeFile(diff.getOldPath());
-            String newPath = ProjectPaths.canonicalizeFile(diff.getNewPath());
-
-            // project configuration change
-            if (ProjectStructure.PROJECT_CONFIG_PATH.equals(oldPath) || ProjectStructure.PROJECT_CONFIG_PATH.equals(newPath))
-            {
-                // technically, we know the only probable operation that can happen is MODIFICATION, CREATE is really an edge case
-                isProjectConfigurationUpdated.set(true);
-                return;
-            }
-
-            EntitySourceDirectory oldPathSourceDirectory = fromProjectStructure.findSourceDirectoryForEntityFilePath(oldPath);
-            EntitySourceDirectory newPathSourceDirectory = toProjectStructure.findSourceDirectoryForEntityFilePath(newPath);
-
-            // entity file change
-            if ((oldPathSourceDirectory != null) || (newPathSourceDirectory != null))
-            {
-                String oldEntityPath = (oldPathSourceDirectory == null) ? oldPath : oldPathSourceDirectory.filePathToEntityPath(oldPath);
-                String newEntityPath = (newPathSourceDirectory == null) ? newPath : newPathSourceDirectory.filePathToEntityPath(newPath);
-                EntityChangeType entityChangeType;
-                if (diff.getDeletedFile())
-                {
-                    entityChangeType = EntityChangeType.DELETE;
-                }
-                else if (diff.getNewFile())
-                {
-                    entityChangeType = EntityChangeType.CREATE;
-                }
-                else if (diff.getRenamedFile())
-                {
-                    entityChangeType = EntityChangeType.RENAME;
-                }
-                else
-                {
-                    entityChangeType = EntityChangeType.MODIFY;
-                }
-                entityDiffs.add(new EntityDiff()
-                {
-                    @Override
-                    public EntityChangeType getEntityChangeType()
-                    {
-                        return entityChangeType;
-                    }
-
-                    @Override
-                    public String getNewPath()
-                    {
-                        return newEntityPath;
-                    }
-
-                    @Override
-                    public String getOldPath()
-                    {
-                        return oldEntityPath;
-                    }
-                });
-            }
-
-            // SKIP non-entity, non-config file
-        });
-        return new Comparison()
-        {
-            @Override
-            public String getToRevisionId()
-            {
-                return toRevisionId;
-            }
-
-            @Override
-            public String getFromRevisionId()
-            {
-                return fromRevisionId;
-            }
-
-            @Override
-            public List<EntityDiff> getEntityDiffs()
-            {
-                return entityDiffs;
-            }
-
-            @Override
-            public boolean isProjectConfigurationUpdated()
-            {
-                return isProjectConfigurationUpdated.get();
-            }
-        };
+        return FileDiff.newFileDiff(diff.getOldPath(), diff.getNewPath(), diff.getDeletedFile(), diff.getNewFile(), diff.getRenamedFile());
     }
 }
