+++ legend-sdlc-server/src/main/java/org/finos/legend/sdlc/server/gitlab/api/GitLabEntityApi.java
@@ -14,14 +14,10 @@
 
 package org.finos.legend.sdlc.server.gitlab.api;
 
-import org.eclipse.collections.api.factory.Lists;
-import org.eclipse.collections.api.factory.Maps;
-import org.eclipse.collections.api.list.MutableList;
-import org.eclipse.collections.api.map.MutableMap;
-import org.eclipse.collections.impl.utility.ListIterate;
+import org.finos.legend.sdlc.core.entity.EntityAccessOperations;
+import org.finos.legend.sdlc.core.entity.EntityModificationOperations;
 import org.finos.legend.sdlc.domain.model.entity.Entity;
 import org.finos.legend.sdlc.domain.model.entity.change.EntityChange;
-import org.finos.legend.sdlc.domain.model.entity.change.EntityChangeType;
 import org.finos.legend.sdlc.domain.model.revision.Revision;
 import org.finos.legend.sdlc.server.domain.api.entity.EntityAccessContext;
 import org.finos.legend.sdlc.server.domain.api.entity.EntityApi;
@@ -33,34 +29,19 @@ import org.finos.legend.sdlc.server.error.LegendSDLCServerException;
 import org.finos.legend.sdlc.server.gitlab.GitLabConfiguration;
 import org.finos.legend.sdlc.server.gitlab.GitLabProjectId;
 import org.finos.legend.sdlc.server.gitlab.auth.GitLabUserContext;
-import org.finos.legend.sdlc.project.structure.EntitySourceDirectory;
-import org.finos.legend.sdlc.project.files.CachingFileAccessContext;
 import org.finos.legend.sdlc.project.files.ProjectFileAccessProvider;
-import org.finos.legend.sdlc.project.files.ProjectFileOperation;
-import org.finos.legend.sdlc.project.structure.ProjectStructure;
 import org.finos.legend.sdlc.server.tools.BackgroundTaskProcessor;
-import org.finos.legend.sdlc.tools.StringTools;
-import org.finos.legend.sdlc.tools.entity.EntityPaths;
 import org.gitlab4j.api.models.DiffRef;
 import org.gitlab4j.api.models.MergeRequest;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import javax.inject.Inject;
-import javax.ws.rs.core.Response.Status;
 
 public class GitLabEntityApi extends GitLabApiWithFileAccess implements EntityApi
 {
-    private static final Logger LOGGER = LoggerFactory.getLogger(GitLabEntityApi.class);
-
     @Inject
     public GitLabEntityApi(GitLabConfiguration gitLabConfiguration, GitLabUserContext userContext, BackgroundTaskProcessor backgroundTaskProcessor)
     {
@@ -162,31 +143,7 @@ public class GitLabEntityApi extends GitLabApiWithFileAccess implements EntityAp
         {
             try
             {
-                ProjectFileAccessProvider.FileAccessContext fileAccessContext = getFileAccessContext(getProjectFileAccessProvider());
-                ProjectStructure projectStructure = ProjectStructure.getProjectStructure(fileAccessContext);
-                for (EntitySourceDirectory sourceDirectory : projectStructure.getEntitySourceDirectories())
-                {
-                    String filePath = sourceDirectory.entityPathToFilePath(path);
-                    ProjectFileAccessProvider.ProjectFile file = fileAccessContext.getFile(filePath);
-                    if (file != null)
-                    {
-                        try
-                        {
-                            Entity localEntity = sourceDirectory.deserialize(file);
-                            if (!Objects.equals(localEntity.getPath(), path))
-                            {
-                                throw new RuntimeException("Expected entity path " + path + ", found " + localEntity.getPath());
-                            }
-                            return localEntity;
-                        }
-                        catch (Exception e)
-                        {
-                            StringBuilder builder = new StringBuilder("Error deserializing entity \"").append(path).append("\" from file \"").append(filePath).append('"');
-                            StringTools.appendThrowableMessageIfPresent(builder, e);
-                            throw new LegendSDLCServerException(builder.toString(), e);
-                        }
-                    }
-                }
+                return EntityAccessOperations.getEntity(getFileAccessContext(getProjectFileAccessProvider()), path, getInfoForException());
             }
             catch (Exception e)
             {
@@ -195,25 +152,14 @@ public class GitLabEntityApi extends GitLabApiWithFileAccess implements EntityAp
                         () -> "Unknown entity " + path + " for " + getInfoForException(),
                         () -> "Failed to get entity " + path + " for " + getInfoForException());
             }
-            throw new LegendSDLCServerException("Unknown entity " + path + " for " + getInfoForException(), Status.NOT_FOUND);
         }
 
         @Override
         public List<Entity> getEntities(Predicate<String> entityPathPredicate, Predicate<String> classifierPathPredicate, Predicate<? super Map<String, ?>> entityContentPredicate, boolean excludeInvalid)
         {
-            try (Stream<EntityProjectFile> stream = getEntityProjectFiles(getFileAccessContext(getProjectFileAccessProvider()), entityPathPredicate, classifierPathPredicate, entityContentPredicate, excludeInvalid))
+            try
             {
-                return stream.map(excludeInvalid ? epf ->
-                {
-                    try
-                    {
-                        return epf.getEntity();
-                    }
-                    catch (Exception ignore)
-                    {
-                        return null;
-                    }
-                } : EntityProjectFile::getEntity).filter(Objects::nonNull).collect(Collectors.toList());
+                return EntityAccessOperations.getEntities(getFileAccessContext(getProjectFileAccessProvider()), entityPathPredicate, classifierPathPredicate, entityContentPredicate, excludeInvalid);
             }
             catch (Exception e)
             {
@@ -227,9 +173,9 @@ public class GitLabEntityApi extends GitLabApiWithFileAccess implements EntityAp
         @Override
         public List<String> getEntityPaths(Predicate<String> entityPathPredicate, Predicate<String> classifierPathPredicate, Predicate<? super Map<String, ?>> entityContentPredicate)
         {
-            try (Stream<EntityProjectFile> stream = getEntityProjectFiles(getFileAccessContext(getProjectFileAccessProvider()), entityPathPredicate, classifierPathPredicate, entityContentPredicate))
+            try
             {
-                return stream.map(EntityProjectFile::getEntityPath).collect(Collectors.toList());
+                return EntityAccessOperations.getEntityPaths(getFileAccessContext(getProjectFileAccessProvider()), entityPathPredicate, classifierPathPredicate, entityContentPredicate);
             }
             catch (Exception e)
             {
@@ -267,7 +213,17 @@ public class GitLabEntityApi extends GitLabApiWithFileAccess implements EntityAp
         {
             LegendSDLCServerException.validateNonNull(entities, "entities may not be null");
             LegendSDLCServerException.validateNonNull(message, "message may not be null");
-            return GitLabEntityApi.this.updateEntities(this.projectId, this.sourceSpecification, entities, replace, message);
+            try
+            {
+                return EntityModificationOperations.updateEntities(getProjectFileAccessProvider(), this.projectId, this.sourceSpecification, entities, replace, message, getReferenceInfo(this.projectId, this.sourceSpecification));
+            }
+            catch (Exception e)
+            {
+                throw buildException(e,
+                        () -> "User " + getCurrentUser() + " is not allowed to perform changes on " + getReferenceInfo(this.projectId, this.sourceSpecification),
+                        () -> "Unknown " + getReferenceInfo(this.projectId, this.sourceSpecification),
+                        () -> "Failed to perform changes on " + getReferenceInfo(this.projectId, this.sourceSpecification) + " (message: " + message + ")");
+            }
         }
 
         @Override
@@ -275,495 +231,18 @@ public class GitLabEntityApi extends GitLabApiWithFileAccess implements EntityAp
         {
             LegendSDLCServerException.validateNonNull(changes, "changes may not be null");
             LegendSDLCServerException.validateNonNull(message, "message may not be null");
-            validateEntityChanges(changes);
-            return GitLabEntityApi.this.performChanges(this.projectId, this.sourceSpecification, revisionId, message, changes);
-        }
-    }
-
-    private Revision updateEntities(String projectId, WorkspaceSourceSpecification sourceSpecification, Iterable<? extends Entity> newEntities, boolean replace, String message)
-    {
-        MutableMap<String, Entity> newEntityDefinitions = indexAndValidateEntitiesForUpdate(newEntities);
-
-        ProjectFileAccessProvider fileProvider = getProjectFileAccessProvider();
-
-        Revision currentWorkspaceRevision = fileProvider.getRevisionAccessContext(projectId, sourceSpecification, null).getCurrentRevision();
-        if (currentWorkspaceRevision == null)
-        {
-            throw new LegendSDLCServerException("Could not find current revision for " + getReferenceInfo(projectId, sourceSpecification) + ": it may be corrupt");
-        }
-        String revisionId = currentWorkspaceRevision.getId();
-        LOGGER.debug("Using revision {} for reference in entity update in {} in project {}", revisionId, sourceSpecification, projectId);
-        List<EntityChange> entityChanges = Lists.mutable.ofInitialCapacity(newEntityDefinitions.size());
-        if (newEntityDefinitions.notEmpty())
-        {
-            try (Stream<EntityProjectFile> stream = getEntityProjectFiles(fileProvider.getFileAccessContext(projectId, sourceSpecification, revisionId)))
+            EntityModificationOperations.validateEntityChanges(changes);
+            try
             {
-                stream.forEach(epf ->
-                {
-                    String path = epf.getEntityPath();
-                    Entity newDefinition = newEntityDefinitions.remove(path);
-                    if (newDefinition != null)
-                    {
-                        Entity entity = epf.getEntity();
-                        String newClassifierPath = newDefinition.getClassifierPath();
-                        Map<String, ?> newContent = newDefinition.getContent();
-                        if (!newClassifierPath.equals(entity.getClassifierPath()) || !newContent.equals(entity.getContent()))
-                        {
-                            entityChanges.add(EntityChange.newModifyEntity(path, newClassifierPath, newContent));
-                        }
-                    }
-                    else if (replace)
-                    {
-                        entityChanges.add(EntityChange.newDeleteEntity(path));
-                    }
-                });
+                return EntityModificationOperations.performChanges(getProjectFileAccessProvider(), this.projectId, this.sourceSpecification, revisionId, message, changes);
             }
-            newEntityDefinitions.forEachValue(definition -> entityChanges.add(EntityChange.newCreateEntity(definition.getPath(), definition.getClassifierPath(), definition.getContent())));
-        }
-        else if (replace)
-        {
-            try (Stream<EntityProjectFile> stream = getEntityProjectFiles(fileProvider.getFileAccessContext(projectId, sourceSpecification, revisionId)))
+            catch (Exception e)
             {
-                stream.map(EntityProjectFile::getEntityPath).map(EntityChange::newDeleteEntity).forEach(entityChanges::add);
+                throw buildException(e,
+                        () -> "User " + getCurrentUser() + " is not allowed to perform changes on " + getReferenceInfo(this.projectId, this.sourceSpecification),
+                        () -> "Unknown " + getReferenceInfo(this.projectId, this.sourceSpecification),
+                        () -> "Failed to perform changes on " + getReferenceInfo(this.projectId, this.sourceSpecification) + " (message: " + message + ")");
             }
-        }
-
-        return performChanges(projectId, sourceSpecification, revisionId, message, entityChanges);
-    }
-
-    private MutableMap<String, Entity> indexAndValidateEntitiesForUpdate(Iterable<? extends Entity> newEntities)
-    {
-        MutableMap<String, Entity> newEntityDefinitions = Maps.mutable.empty();
-        MutableList<String> errorMessages = Lists.mutable.empty();
-        newEntities.forEach(entity ->
-        {
-            if (entity == null)
-            {
-                errorMessages.add("Invalid entity: null");
-                return;
-            }
-
-            String path = entity.getPath();
-            if (!EntityPaths.isValidEntityPath(path))
-            {
-                errorMessages.add("Invalid entity path: \"" + path + "\"");
-                return;
-            }
-
-            String classifierPath = entity.getClassifierPath();
-            if (!EntityPaths.isValidClassifierPath(classifierPath))
-            {
-                errorMessages.add("Entity: " + path + "; error: invalid classifier path \"" + classifierPath + "\"");
-            }
-
-            Map<String, ?> content = entity.getContent();
-            if (content == null)
-            {
-                errorMessages.add("Entity: " + path + "; error: missing content");
-            }
-            else
-            {
-                Object pkg = content.get("package");
-                Object name = content.get("name");
-                if (!(pkg instanceof String) || !(name instanceof String) || !path.equals(pkg + EntityPaths.PACKAGE_SEPARATOR + name))
-                {
-                    StringBuilder builder = new StringBuilder("Entity: ").append(path).append("; mismatch between entity path and package (");
-                    if (pkg instanceof String)
-                    {
-                        builder.append('"').append(pkg).append('"');
-                    }
-                    else
-                    {
-                        builder.append(pkg);
-                    }
-                    builder.append(") and name (");
-                    if (name instanceof String)
-                    {
-                        builder.append('"').append(name).append('"');
-                    }
-                    else
-                    {
-                        builder.append(name);
-                    }
-                    errorMessages.add(builder.append(") properties").toString());
-                }
-            }
-
-            Entity oldDefinition = newEntityDefinitions.put(path, entity);
-            if (oldDefinition != null)
-            {
-                errorMessages.add("Entity: " + path + "; error: multiple definitions");
-            }
-        });
-        if (errorMessages.notEmpty())
-        {
-            throw new LegendSDLCServerException((errorMessages.size() == 1) ? errorMessages.get(0) : errorMessages.makeString("There are errors with entity definitions:\n\t", "\n\t", ""), Status.BAD_REQUEST);
-        }
-        return newEntityDefinitions;
-    }
-
-    private Revision performChanges(String projectId, WorkspaceSourceSpecification sourceSpecification, String referenceRevisionId, String message, List<? extends EntityChange> changes)
-    {
-        int changeCount = changes.size();
-        if (changeCount == 0)
-        {
-            LOGGER.debug("No changes for {} in project {}", sourceSpecification, projectId);
-            return null;
-        }
-        LOGGER.debug("Committing {} changes to {} in project {}: {}", changeCount, sourceSpecification, projectId, message);
-        try
-        {
-            ProjectFileAccessProvider.FileAccessContext fileAccessContext = getProjectFileAccessProvider().getFileAccessContext(projectId, sourceSpecification, referenceRevisionId);
-            ProjectStructure projectStructure = ProjectStructure.getProjectStructure(fileAccessContext);
-            MutableList<ProjectFileOperation> fileOperations = ListIterate.collect(changes, c -> entityChangeToFileOperation(c, projectStructure, fileAccessContext));
-            fileOperations.removeIf(Objects::isNull);
-            if (fileOperations.isEmpty())
-            {
-                LOGGER.debug("No changes for {} in project {}", sourceSpecification, projectId);
-                return null;
-            }
-            return getProjectFileAccessProvider().getFileModificationContext(projectId, sourceSpecification, referenceRevisionId).submit(message, fileOperations);
-        }
-        catch (Exception e)
-        {
-            throw buildException(e,
-                    () -> "User " + getCurrentUser() + " is not allowed to perform changes on " + getReferenceInfo(projectId, sourceSpecification),
-                    () -> "Unknown " + getReferenceInfo(projectId, sourceSpecification),
-                    () -> "Failed to perform changes on "  + getReferenceInfo(projectId, sourceSpecification) + " (message: " + message + ")");
-        }
-    }
-
-    private ProjectFileOperation entityChangeToFileOperation(EntityChange change, ProjectStructure projectStructure, ProjectFileAccessProvider.FileAccessContext fileAccessContext)
-    {
-        switch (change.getType())
-        {
-            case CREATE:
-            {
-                String entityPath = change.getEntityPath();
-
-                // check if a file already exists for this entity
-                if (projectStructure.findEntityFile(entityPath, fileAccessContext) != null)
-                {
-                    throw new LegendSDLCServerException("Unable to handle operation " + change + ": entity \"" + entityPath + "\" already exists");
-                }
-
-                Entity entity = Entity.newEntity(entityPath, change.getClassifierPath(), change.getContent());
-                EntitySourceDirectory sourceDirectory = projectStructure.findSourceDirectoryForEntity(entity);
-                if (sourceDirectory == null)
-                {
-                    throw new LegendSDLCServerException("Unable to handle operation " + change + ": cannot serialize entity \"" + entityPath + "\"");
-                }
-                return ProjectFileOperation.addFile(sourceDirectory.entityPathToFilePath(change.getEntityPath()), sourceDirectory.serializeToBytes(entity));
-            }
-            case DELETE:
-            {
-                String entityPath = change.getEntityPath();
-                String filePath = projectStructure.findEntityFile(entityPath, fileAccessContext);
-                if (filePath == null)
-                {
-                    throw new LegendSDLCServerException("Unable to handle operation " + change + ": could not find entity \"" + entityPath + "\"");
-                }
-                return ProjectFileOperation.deleteFile(filePath);
-            }
-            case MODIFY:
-            {
-                String entityPath = change.getEntityPath();
-                Entity entity = Entity.newEntity(entityPath, change.getClassifierPath(), change.getContent());
-
-                // find current file
-                String currentFilePath = projectStructure.findEntityFile(entityPath, fileAccessContext);
-                if (currentFilePath == null)
-                {
-                    throw new LegendSDLCServerException("Unable to handle operation " + change + ": could not find entity \"" + entityPath + "\"");
-                }
-
-                EntitySourceDirectory newSourceDirectory = projectStructure.findSourceDirectoryForEntity(entity);
-                if (newSourceDirectory == null)
-                {
-                    throw new LegendSDLCServerException("Unable to handle operation " + change + ": cannot serialize entity \"" + entityPath + "\"");
-                }
-
-                String newFilePath = newSourceDirectory.entityPathToFilePath(entityPath);
-                byte[] serialized = newSourceDirectory.serializeToBytes(entity);
-
-                if (!currentFilePath.equals(newFilePath))
-                {
-                    return ProjectFileOperation.moveFile(currentFilePath, newFilePath, serialized);
-                }
-                if (!Arrays.equals(serialized, fileAccessContext.getFile(currentFilePath).getContentAsBytes()))
-                {
-                    return ProjectFileOperation.modifyFile(currentFilePath, serialized);
-                }
-
-                return null;
-            }
-            case RENAME:
-            {
-                String currentEntityPath = change.getEntityPath();
-                for (EntitySourceDirectory sourceDirectory : projectStructure.getEntitySourceDirectories())
-                {
-                    String filePath = sourceDirectory.entityPathToFilePath(currentEntityPath);
-                    if (fileAccessContext.fileExists(filePath))
-                    {
-                        String newFilePath = sourceDirectory.entityPathToFilePath(change.getNewEntityPath());
-                        return ProjectFileOperation.moveFile(filePath, newFilePath);
-                    }
-                }
-                throw new LegendSDLCServerException("Unable to handle operation " + change + ": could not find entity \"" + currentEntityPath + "\"");
-            }
-            default:
-            {
-                throw new RuntimeException("Unknown entity change type: " + change.getType());
-            }
-        }
-    }
-
-    private Stream<EntityProjectFile> getEntityProjectFiles(ProjectFileAccessProvider.FileAccessContext accessContext, Predicate<String> entityPathPredicate, Predicate<String> classifierPathPredicate, Predicate<? super Map<String, ?>> contentPredicate)
-    {
-        return getEntityProjectFiles(accessContext, entityPathPredicate, classifierPathPredicate, contentPredicate, false);
-    }
-
-    private Stream<EntityProjectFile> getEntityProjectFiles(ProjectFileAccessProvider.FileAccessContext accessContext, Predicate<String> entityPathPredicate, Predicate<String> classifierPathPredicate, Predicate<? super Map<String, ?>> contentPredicate, boolean excludeInvalid)
-    {
-        Stream<EntityProjectFile> stream = getEntityProjectFiles(accessContext);
-        if (entityPathPredicate != null)
-        {
-            stream = stream.filter(epf -> entityPathPredicate.test(epf.getEntityPath()));
-        }
-        if (classifierPathPredicate != null)
-        {
-            stream = stream.filter(excludeInvalid ? epf ->
-            {
-                Entity entity;
-                try
-                {
-                    entity =  epf.getEntity();
-                }
-                catch (Exception ignore)
-                {
-                    return false;
-                }
-                return classifierPathPredicate.test(entity.getClassifierPath());
-            } : epf -> classifierPathPredicate.test(epf.getEntity().getClassifierPath()));
-        }
-        if (contentPredicate != null)
-        {
-            stream = stream.filter(excludeInvalid ? epf ->
-            {
-                Entity entity;
-                try
-                {
-                    entity =  epf.getEntity();
-                }
-                catch (Exception ignore)
-                {
-                    return false;
-                }
-                return contentPredicate.test(entity.getContent());
-            } : epf -> contentPredicate.test(epf.getEntity().getContent()));
-        }
-        return stream;
-    }
-
-    private Stream<EntityProjectFile> getEntityProjectFiles(ProjectFileAccessProvider.FileAccessContext accessContext)
-    {
-        ProjectStructure projectStructure = ProjectStructure.getProjectStructure(accessContext);
-        List<EntitySourceDirectory> sourceDirectories = projectStructure.getEntitySourceDirectories();
-        ProjectFileAccessProvider.FileAccessContext cachingAccessContext = (sourceDirectories.size() > 1) ? CachingFileAccessContext.wrap(accessContext) : accessContext;
-        return sourceDirectories.stream().flatMap(sd -> getSourceDirectoryProjectFiles(cachingAccessContext, sd));
-    }
-
-    private Stream<EntityProjectFile> getSourceDirectoryProjectFiles(ProjectFileAccessProvider.FileAccessContext accessContext, EntitySourceDirectory sourceDirectory)
-    {
-        return accessContext.getFilesInDirectory(sourceDirectory.getDirectory())
-                .filter(f -> sourceDirectory.isPossiblyEntityFilePath(f.getPath()))
-                .map(f -> new EntityProjectFile(sourceDirectory, f));
-    }
-
-    private static void validateEntityChanges(List<? extends EntityChange> entityChanges)
-    {
-        StringBuilder builder = new StringBuilder();
-        List<String> errorMessages = Lists.mutable.ofInitialCapacity(4);
-        int i = 0;
-        for (EntityChange change : entityChanges)
-        {
-            collectErrorsForEntityChange(change, errorMessages);
-            if (!errorMessages.isEmpty())
-            {
-                if (builder.length() == 0)
-                {
-                    builder.append("There are entity change errors:");
-                }
-                builder.append("\n\tEntity change #").append(i + 1).append(" (").append(change).append("):");
-                errorMessages.forEach(m -> builder.append("\n\t\t").append(m));
-                errorMessages.clear();
-            }
-            i++;
-        }
-        if (builder.length() > 0)
-        {
-            throw new LegendSDLCServerException(builder.toString(), Status.BAD_REQUEST);
-        }
-    }
-
-    private static void collectErrorsForEntityChange(EntityChange entityChange, Collection<? super String> errorMessages)
-    {
-        if (entityChange == null)
-        {
-            errorMessages.add("Invalid entity change: null");
-            return;
-        }
-
-        EntityChangeType type = entityChange.getType();
-        if (type == null)
-        {
-            errorMessages.add("Missing entity change type");
-        }
-
-        String path = entityChange.getEntityPath();
-        String classifierPath = entityChange.getClassifierPath();
-        Map<String, ?> content = entityChange.getContent();
-        String newPath = entityChange.getNewEntityPath();
-
-        if (path == null)
-        {
-            errorMessages.add("Missing entity path");
-        }
-        else if (!EntityPaths.isValidEntityPath(path))
-        {
-            errorMessages.add("Invalid entity path: " + path);
-        }
-        else if (content != null)
-        {
-            Object pkg = content.get("package");
-            Object name = content.get("name");
-            if (!(pkg instanceof String) || !(name instanceof String) || !path.equals(pkg + EntityPaths.PACKAGE_SEPARATOR + name))
-            {
-                StringBuilder builder = new StringBuilder("Mismatch between entity path (\"").append(path).append("\") and package (");
-                if (pkg instanceof String)
-                {
-                    builder.append('"').append(pkg).append('"');
-                }
-                else
-                {
-                    builder.append(pkg);
-                }
-                builder.append(") and name (");
-                if (name instanceof String)
-                {
-                    builder.append('"').append(name).append('"');
-                }
-                else
-                {
-                    builder.append(name);
-                }
-                builder.append(") properties");
-                errorMessages.add(builder.toString());
-            }
-        }
-        if (type != null)
-        {
-            switch (type)
-            {
-                case CREATE:
-                case MODIFY:
-                {
-                    if (classifierPath == null)
-                    {
-                        errorMessages.add("Missing classifier path");
-                    }
-                    else if (!EntityPaths.isValidClassifierPath(classifierPath))
-                    {
-                        errorMessages.add("Invalid classifier path: " + classifierPath);
-                    }
-                    if (content == null)
-                    {
-                        errorMessages.add("Missing content");
-                    }
-                    if (newPath != null)
-                    {
-                        errorMessages.add("Unexpected new entity path: " + newPath);
-                    }
-                    break;
-                }
-                case RENAME:
-                {
-                    if (classifierPath != null)
-                    {
-                        errorMessages.add("Unexpected classifier path: " + classifierPath);
-                    }
-                    if (content != null)
-                    {
-                        errorMessages.add("Unexpected content");
-                    }
-                    if (newPath == null)
-                    {
-                        errorMessages.add("Missing new entity path");
-                    }
-                    else if (!EntityPaths.isValidEntityPath(newPath))
-                    {
-                        errorMessages.add("Invalid new entity path: " + newPath);
-                    }
-                    break;
-                }
-                case DELETE:
-                {
-                    if (classifierPath != null)
-                    {
-                        errorMessages.add("Unexpected classifier path: " + classifierPath);
-                    }
-                    if (content != null)
-                    {
-                        errorMessages.add("Unexpected content");
-                    }
-                    if (newPath != null)
-                    {
-                        errorMessages.add("Unexpected new entity path: " + newPath);
-                    }
-                    break;
-                }
-                default:
-                {
-                    errorMessages.add("Unexpected entity change type: " + type);
-                }
-            }
-        }
-    }
-
-    private static class EntityProjectFile
-    {
-        private final EntitySourceDirectory sourceDirectory;
-        private final ProjectFileAccessProvider.ProjectFile file;
-        private String path;
-        private Entity entity;
-
-        private EntityProjectFile(EntitySourceDirectory sourceDirectory, ProjectFileAccessProvider.ProjectFile file)
-        {
-            this.sourceDirectory = sourceDirectory;
-            this.file = file;
-        }
-
-        synchronized String getEntityPath()
-        {
-            if (this.path == null)
-            {
-                this.path = this.sourceDirectory.filePathToEntityPath(this.file.getPath());
-            }
-            return this.path;
-        }
-
-        synchronized Entity getEntity()
-        {
-            if (this.entity == null)
-            {
-                Entity localEntity = this.sourceDirectory.deserialize(this.file);
-                if (!Objects.equals(localEntity.getPath(), getEntityPath()))
-                {
-                    throw new RuntimeException("Expected entity path " + getEntityPath() + ", found " + localEntity.getPath());
-                }
-                this.entity = localEntity;
-            }
-            return this.entity;
         }
     }
 }
